Initial working prototype
Runs a webserver on a custom port
Then a PUT is received, the Host is extracted and the request is forwarded to Host

The request and response are pushed into a queue

A separate thread reads the queue. if a handler is set for the url invoked, the handler is called with all data

A handler for goodwe is implemented, decoding the binary data and logging it.